Improve HashLink executable file system exists check with Lime >= 8.0 (closes #110)
The executable will be created as part of the preLaunchTask, so it may not exist yet unless there was a previous successful build.

After the preLaunchTask, it will exist. However, HashLink debugger will report an error if it doesn't, so this check is redundant. On the other hand, I decided to keep the check for older Lime. Since we're using the executable that's bundled with Lime, our error will include the Lime version and expected path, which can help with debugging. For Lime >= 8.0, the path is always within the project export folder, and version matters less.